+++ backend/main/resources/User.py
@@ -5,7 +5,7 @@ from main.models import UserModel
 from main.models import PoemModel
 from main.models import ReviewModel
 from sqlalchemy import func
-from flask_jwt_extended import jwt_required, get_jwt_identity
+from flask_jwt_extended import jwt_required, get_jwt_identity, get_jwt
 from main.auth.decorators import admin_required
 
 #Recurso usuario
@@ -15,25 +15,40 @@ class User(Resource):
     @jwt_required(optional=True)
     def get(self, id):
         user = db.session.query(UserModel).get_or_404(id)
-        return user.to_json()
+        token_id = get_jwt_identity()
+        claims = get_jwt()
+        if not claims:
+            claims = {'admin': False}
+        if token_id == user.id or claims['admin']:
+            return user.to_json()
+        else:
+            return user.to_json_short()
 
     #eliminar recurso
-    # @admin_required
+    @admin_required
     def delete(self, id):
         user = db.session.query(UserModel).get_or_404(id)
-        db.session.delete(user)
-        db.session.commit()
-        return 'User deleted', 204
+        if not user.admin:
+            db.session.delete(user)
+            db.session.commit()
+            return 'User deleted', 204
+        else:
+            return 'Admin User can not be deleted', 403
     
+    #modificar recurso
     @jwt_required()
     def put(self, id):
         user = db.session.query(UserModel).get_or_404(id)
         data = request.get_json().items()
-        for key,value in data:
-            setattr(user,key,value)
-        db.session.add(user)
-        db.session.commit()
-        return user.to_json(), 201
+        token_id = get_jwt_identity()
+        if token_id == user.id:
+            for key,value in data:
+                setattr(user,key,value)
+            db.session.add(user)
+            db.session.commit()
+            return user.to_json(), 201
+        else:
+            return 'Not allowed, only owner can modify', 403
 
 class Users(Resource):
 
@@ -66,7 +81,6 @@ class Users(Resource):
                     if value == 'cant_poem':
                         users = users.outerjoin(UserModel.poems).group_by(UserModel.id).order_by(func.count(UserModel.poems))
 
-                        # professors=professors.outerjoin(ProfessorModel.projects).group_by(ProfessorModel.id).order_by(func.count(ProjectModel.id).desc())
         #Ahora pagino, guarde la consulta parcial en users
         users = users.paginate(page, per_page, True, 20) #Ahora no es una lista de lementos, es una paginacion
         return jsonify({"users":[user.to_json_short() for user in users.items],
